+++ libvg/vg_info.c
@@ -22,20 +22,20 @@ const t_str	g_vg_contributors[5] = {
 
 void	vg_icon(void)
 {
-	printf("\e[94m\e[48;5;235m     / /  |/| |             \e[48;5;235m\e[0m\n");
-	printf("\e[94m\e[48;5;235m   / /__   / /              \e[48;5;235m\e[0m\n");
-	printf("\e[94m\e[48;5;235m /____  | | |/|  █  Vogship \e[48;5;235m\e[0m\n");
-	printf("\e[94m\e[48;5;235m      | |                   \e[48;5;235m\e[0m\n");
+	ft_printf("\e[94m\e[48;5;235m     / /  |/| |             \e[48;5;235m\e[0m\n");
+	ft_printf("\e[94m\e[48;5;235m   / /__   / /              \e[48;5;235m\e[0m\n");
+	ft_printf("\e[94m\e[48;5;235m /____  | | |/|  █  Vogship \e[48;5;235m\e[0m\n");
+	ft_printf("\e[94m\e[48;5;235m      | |                   \e[48;5;235m\e[0m\n");
 }
 
 void	vg_version(void)
 {
-	printf("\e[92mVersion %s\e[0m\n", VG_VER);
+	ft_printf("\e[92mVersion %s\e[0m\n", VG_VER);
 }
 
 void	vg_author(void)
 {
-	printf("\e[96mWritten by %s\e[0m\n", VG_AUTHOR);
+	ft_printf("\e[96mWritten by %s\e[0m\n", VG_AUTHOR);
 }
 
 void	vg_contributors(void)
@@ -43,7 +43,7 @@ void	vg_contributors(void)
 	int	i;
 
 	i = 0;
-	printf("\e[96mContributions from:\e[0m\n");
+	ft_printf("\e[96mContributions from:\e[0m\n");
 	while (g_vg_contributors[i])
-		printf("    %s\n", g_vg_contributors[i++]);
+		ft_printf("    %s\n", g_vg_contributors[i++]);
 }